chore: moved block signing to common
