+++ src/signers/aws_kms.rs
@@ -1,8 +1,10 @@
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
-use alloy::network::EthereumWallet;
-use anyhow::{anyhow, Result as AnyhowResult};
-use alloy::signers::{aws::AwsSigner, Signer};
+use anyhow::{Result as AnyhowResult};
+use alloy::{
+    primitives::{Address},
+    signers::{aws::AwsSigner, Signer},
+};
 use aws_config::BehaviorVersion;
 use aws_sdk_kms::Client;
 use axum::http::StatusCode;
@@ -23,13 +25,11 @@ use tracing::info;
 
 use crate::jsonrpc::AddressResponse;
 use crate::{
-    app_types::{AppError, AppJson, AppResult},
-    jsonrpc::{JsonRpcReply, JsonRpcRequest, JsonRpcResult},
+    app_types::{AppJson, AppResult},
+    jsonrpc::{JsonRpcReply, JsonRpcRequest},
     shutdown_signal::shutdown_signal,
-    signers::common::{handle_eth_sign_transaction, handle_health_status, to_signing_hash, BlockPayloadArgs},
+    signers::common::{handle_eth_sign_jsonrpc},
 };
-use::alloy::hex;
-use alloy::primitives::{Address};
 
 
 #[derive(StructOpt)]
@@ -46,7 +46,7 @@ pub enum AwsCommand {
 #[derive(Clone)]
 struct AppState {
     client: Client,
-    signers: Arc<Mutex<HashMap<String, AwsSigner>>>,
+    signers: Arc<Mutex<HashMap<String, Arc<AwsSigner>>>>,
 }
 
 const API_TIMEOUT_SECS: u64 = 30;
@@ -66,14 +66,14 @@ async fn handle_request(
     handle_eth_sign_jsonrpc(payload, signer).await
 }
 
-async fn get_signer(state: Arc<AppState>, key_id: String) -> AnyhowResult<AwsSigner> {
+async fn get_signer(state: Arc<AppState>, key_id: String) -> AnyhowResult<Arc<AwsSigner>> {
     let mut signers = state.signers.lock().await;
 
     if let Some(signer) = signers.get(&key_id) {
         return Ok(signer.clone());
     }
 
-    let signer = AwsSigner::new(state.client.clone(), key_id.clone(), None).await?;
+    let signer = Arc::new(AwsSigner::new(state.client.clone(), key_id.clone(), None).await?);
     signers.insert(key_id.clone(), signer.clone());
     Ok(signer)
 }
@@ -132,64 +132,6 @@ pub async fn handle_aws_kms(opt: AwsOpt) {
 }
 
 
-pub async fn handle_eth_sign_jsonrpc(
-    payload: JsonRpcRequest<Vec<Value>>,
-    signer: AwsSigner,
-) -> AppResult<JsonRpcReply<Value>> {
-    let method = payload.method.as_str();
-
-    let result = match method {
-        "eth_signTransaction" => handle_eth_sign_transaction(payload, EthereumWallet::from(signer)).await,
-        "health_status" => handle_health_status(payload).await,
-        "opsigner_signBlockPayload" => handle_eth_sign_block(payload, signer).await,
-        _ => Err(anyhow!(
-            "method not supported (only eth_signTransaction, health_status and opsigner_signBlockPayload): {}",
-            method
-        )),
-    };
-
-    result.map(AppJson).map_err(AppError)
-}
-
-
-
-
-pub async fn handle_eth_sign_block(
-    payload: JsonRpcRequest<Vec<Value>>,
-    signer: AwsSigner,
-) -> AnyhowResult<JsonRpcReply<Value>> {
-
-    println!("handle_eth_sign_block payload: {:?}", payload);
-    let params = payload.params.ok_or_else(|| anyhow!("params is empty"))?;
-    if params.is_empty() {
-        return Err(anyhow!("params is empty"));
-    }
-
-    let block_object = params[0].clone();
-    let block: BlockPayloadArgs = serde_json::from_value(block_object)?;
-
-    println!("block: {:?}", block);
-    let signing_hash = to_signing_hash(&block);
-
-    let signed_hash  = signer.sign_hash(&signing_hash).await?;
-
-    // extract the 65-byte array
-    let mut sig_bytes: [u8; 65] = signed_hash.as_bytes();
-    if sig_bytes[64] < 27 {
-        return Err(anyhow!("Invalid recovery id: expected value >= 27, got {}", sig_bytes[64]));
-    }
-
-    sig_bytes[64] = sig_bytes[64] - 27; // Adjust the recovery id to be 0 or 1
-
-    // encode as a "0x"-prefixed hex string
-    let signed_hash_hex = hex::encode_prefixed(&sig_bytes[..]);
-    println!("signed_hash_hex: {:?}", signed_hash_hex);
-    Ok(JsonRpcReply {
-        id: payload.id,
-        jsonrpc: payload.jsonrpc,
-        result: JsonRpcResult::Result(Value::String(signed_hash_hex)),
-    })
-}
